requirements.txt
django==5.0.2
django-allauth==0.61.1
djangorestframework==3.14.0
dj-rest-auth==5.0.2
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
jsonschema==4.21.1
phonenumbers==8.13.31
whitenoise==6.6.0
gunicorn==21.2.0